requirements.txt
python-telegram-bot==21.0
requests==2.31.0
python-dotenv==1.0.0
watchdog==3.0.0
